Add generic NJointBlmcRobotDriver and remove calibrate from constructor (#6)
* Add generic NJointBlmcRobotDriver

Generic, templated `NJointBlmcRobotDriver` which can be used as a base
for robots that consist of n joints.

Use this new class as a base for `RealFingerDriver`.

Add helper structs `MotorParameters` and `CalibrationParameters` to
provide them more easily from the derived class.

* Ensure calibrate() is called before applying actions

Add a `is_calibrated_` flag and throw an error if `apply_action` is
called while this flag is false.

* Sharpen controller for calibration

...and reduce the tolerance.

* Add a demo script to just print current position

Initialises and calibrates the robot.  Then zero-torque commands are
send and the position is printed continuously.

* Call calibrate() in demo_real_finger.py

* Auto-reformat

* Pause after calibration

This is needed as the user code may not immediately continue sending
commands after the calibration.  Without the pause, the timeout watchdog
of the micro controller will trigger and go to fault state.

* Rename and cleanup the print position script

* Move create_motors to derived class & refactor c'tors

As the `create_motors` method is highly robot-specific, it should be in
the specific robot class, not the `NJointBlmcRobotDriver`.  Change the
constructors accordingly.

* Set more parameters via constructor

* Add apply_action_uncalibrated for use in homing

Instead of setting `is_calibrated_` before homing to be able to move,
add a `apply_action_uncalibrated` method that does not check
`is_calibrated_`.

* Cleanup some comments

* Rename calibrate to initialize