Cleanup demo_real_finger.py and add comments
Clean up and add comments to the demo_real_finger.py as well.
Remove the threading code as it just adds unnecessary complexity for the
sake of this demo.

It is now mostly identical to demo_fake_finger.py, only difference is
using the real finger backend.

Rename the backend variables in both demos to be explicit about their
nature.