Increase STOP_VELOCITY and add debug output
The old stop velocity threshold was a bit too low.  At the end step each
joint is sometimes above and sometimes below the threshold.
For the single finger this seems okay because there is a high enough
probability that at one timestep all three joints are below the
threshold at the same time.  However, with the nine joints of the
TriFinger this rarely ever happens.

Increasing the threshold to `0.005` is working reliably and should still
be small enough.