Run initialisation in real-time thread
When just calling `initialize()` directly from the user-code, it will
not be executed with real-time guarantees (assuming that the code where
the user calls it is not running in real-time.
This causes non-deterministic delays which can impact control stability
and may trigger the CAN receive timeout on the motor board.
To avoid this issue, simply spawn a real-time thread inside the
`initialize()` method and do the actual initialisation in this thread.