Print travelled distance at end of homing
When the joints are moved to the desired zero position before starting
homing, this corresponds to the home offset.  As the "travelled
distance" it is independent of the configured home offset, this allows
to get a corrected home offset more easily without the need to first
modify the config file.