Use minimum jerk trajectory in move_to_position()
Instead of directly setting the goal position, move more smoothly on a
trajectory that minimizes jerk.

This has several advantages:
 - less stress for the mechanics
 - lower power consumption(?)
 - automatically slows down towards the goal, so no need to explicitly
   stop the robot once the goal is reached

With this, there is no need for a timeout anymore.  Instead a number of
time steps for the motion needs to be passed (the smaller this number,
the faster the motion).