NJointBlmcRobotDriver: Configurable "Shutdown Trajectory"
Add optional parameter "shutdown_trajectory" to the configuration of the
driver.  If set, this parameter is expected to contain a list of target
positions.  During `shutdown()`, the robot is moved to these positions
one by one, using `move_to_position()`.
This can be used to move the robot to a safe rest position before
turning it of.

Since this also needs the `position_tolerance_rad` parameter to
determine if the goal is reached, move this one out of `calibration` and
rename it to `move_to_position_tolerance_rad`.