chore: Add GitHub workflows for build and documentation
`build_pr.yml` runs the build and unit tests on pull requests (to
verify the proposed change do not break something).

`build_master.yml` does the same when something is pushed/merged into
master but additionally it also build the documentation and deploys it
to the GitHub Pages of the repo.