Don't use slider box if port is empty or "none"
Currently the slider box must always be connected, even if not used,
otherwise the robot doesn't run.  Since often it is not actually needed,
change this to only use the slider box if the serial port is set to a
non-empty value unequal "none".

Also refactor a bit: `slider_positions_vector_` (now renamed to
`slider_box_values`) is only used locally and written to in the
beginning, so no need to keep it as class member.