Remove code duplication between Solo8 and Solo12
Apart from some details during initialisation, the code of Solo8 and
Solo12 is mostly the same.  To get rid of this redundancy add a
templated SoloX class (based on Solo12) and simply create typedefs for
Solo8 and Solo12.

For the parts of the initialisation code that differs, add private
methods with explicit specialisations for the two versions.

BREAKING CHANGE: While Solo8 was mostly identical with Solo12, there
where some minor differences.  Code relying on these parts will break by
this change.  However, it should be relatively easy to fix these issues
if they emerge.